[bench] add mechanism to do rmw-rd pattern, effectively disable direct migration

+++ src/utils/lib_mem_region.cc
@@ -7,9 +7,9 @@
 namespace utils {
 
 MemRegion::MemRegion(
-        uint32_t size,
-        uint32_t page_size,
-        uint32_t line_size) :
+        uint64_t size,
+        uint64_t page_size,
+        uint64_t line_size) :
     size_ (size),
     page_size_ (page_size),
     line_size_ (line_size),
@@ -27,17 +27,17 @@ MemRegion::MemRegion(
     srand(0);
 }
 
-void MemRegion::randomizeSequence_(std::vector<uint32_t>& sequence, uint32_t size, uint32_t unit)
+void MemRegion::randomizeSequence_(std::vector<uint64_t>& sequence, uint64_t size, uint64_t unit)
 {
     // initialize to sequential pattern
-    for (uint32_t i = 0; i < size; ++i) {
+    for (uint64_t i = 0; i < size; ++i) {
         sequence[i] = i * unit;
     }
     // randomize by swapping
-    uint32_t r = rand() ^ (rand() << 10);
-    for (uint32_t i = size-1; i> 0; --i) {
+    uint64_t r = rand() ^ (rand() << 10);
+    for (uint64_t i = size-1; i> 0; --i) {
         r = (r << 1) ^ rand();
-        uint32_t tmp = sequence[r % (i+1)];
+        uint64_t tmp = sequence[r % (i+1)];
         sequence[r % (i+1)] = sequence[i];
         sequence[i] = tmp;
     }
@@ -47,7 +47,7 @@ void MemRegion::randomizeSequence_(std::vector<uint32_t>& sequence, uint32_t siz
 void MemRegion::stride_init()
 {
     char* addr = base_;
-    uint32_t i;
+    uint64_t i;
     for (i = line_size_; i < size_; i += line_size_) {
         *(char **)(&addr[i - line_size_]) = (char*)&addr[i];
     }
@@ -57,19 +57,19 @@ void MemRegion::stride_init()
 // create a circular list of pointers with random-in-page
 void MemRegion::page_random_init()
 {
-    std::vector<uint32_t> pages_(num_pages_, 0);
-    std::vector<uint32_t> linesInPage_(num_lines_in_page_, 0);
+    std::vector<uint64_t> pages_(num_pages_, 0);
+    std::vector<uint64_t> linesInPage_(num_lines_in_page_, 0);
     randomizeSequence_(pages_, num_pages_, page_size_);
     randomizeSequence_(linesInPage_, num_lines_in_page_, line_size_);
     // run through the pages
-    for (uint32_t i = 0; i < num_pages_; ++i) {
+    for (uint64_t i = 0; i < num_pages_; ++i) {
         // run through the lines within a page
-        for (uint32_t j = 0; j < num_lines_in_page_ - 1; ++j) {
+        for (uint64_t j = 0; j < num_lines_in_page_ - 1; ++j) {
             *(char**)(base_ + pages_.at(i) + linesInPage_.at(j))
                 = base_ + pages_.at(i) + linesInPage_.at(j+1);
         }
         // jump the next page
-        uint32_t next_page = (i == num_pages_ - 1) ? 0 : (i + 1);
+        uint64_t next_page = (i == num_pages_ - 1) ? 0 : (i + 1);
         *(char**)(base_ + pages_.at(i) + linesInPage_.at(num_lines_in_page_ - 1))
             = base_ + pages_.at(next_page) + linesInPage_.at(0);
     }
@@ -78,11 +78,11 @@ void MemRegion::page_random_init()
 // create a circular list of pointers with all-random
 void MemRegion::all_random_init()
 {
-    const uint32_t num_lines = numLines();
-    std::vector<uint32_t> lines_(num_lines, 0);
+    const uint64_t num_lines = numLines();
+    std::vector<uint64_t> lines_(num_lines, 0);
     randomizeSequence_(lines_, num_lines, line_size_);
     // run through the lines
-    for (uint32_t i = 0; i < num_lines - 1; ++i) {
+    for (uint64_t i = 0; i < num_lines - 1; ++i) {
         *(char**)(base_ + lines_.at(i)) = base_ + lines_.at(i+1);
     }
     *(char**)(base_ + lines_.at(num_lines - 1)) = base_ + lines_.at(0);
@@ -93,7 +93,7 @@ void MemRegion::dump()
     std::cout << "================================" << std::endl;
     std::cout << "size=" << size_ << ", page=" << page_size_ << ", line=" << line_size_
         << ", numPage=" << num_pages_ << ", numLinesInPage=" << num_lines_in_page_ << std::endl;
-    for (uint32_t i = 0; i < size_; i += line_size_) {
+    for (uint64_t i = 0; i < size_; i += line_size_) {
         uint64_t curr = reinterpret_cast<uint64_t>(base_ + i);
         uint64_t next = reinterpret_cast<uint64_t>(*(char**)(base_ + i));
         uint64_t curr_offset = (curr - reinterpret_cast<uint64_t>((char**)base_)) / line_size_;
@@ -104,7 +104,7 @@ void MemRegion::dump()
     }
     std::cout << "--------------------------------" << std::endl;
     char** p = (char**)base_;
-    for (uint32_t i = 0; i < numLines(); ++i) {
+    for (uint64_t i = 0; i < numLines(); ++i) {
         uint64_t curr = reinterpret_cast<uint64_t>(p);
         uint64_t next = reinterpret_cast<uint64_t>(*p);
         uint64_t curr_offset = (curr - reinterpret_cast<uint64_t>((char**)base_)) / line_size_;
